This Grace
In this project, I used HTML, CSS, and JavaScript (jQuery) to create a single-page website for a restaurant.
Here I tried to create a clean and objective website, focused on customer acquisition and product sales.
In this version, the website is optimized for tablets and smartphones. Useful links have been added to the restaurant's contact details and for customer delivery orders.